+++ game/src/main/java/othello/app/Player.java
@@ -1,0 +1,11 @@
+package othello.game;
+
+import othello.*;
+
+import java.util.List;
+
+public interface Player {
+    String getName();
+
+    Field bestMove(Position pos, List<Field> moves);
+}